Move encoders to separate directory
