+++ march_hardware/src/encoder/Encoder.cpp
@@ -1,5 +1,5 @@
 // Copyright 2019 Project March.
-#include "march_hardware/Encoder.h"
+#include "march_hardware/encoder/Encoder.h"
 #include "march_hardware/EtherCAT/EthercatIO.h"
 #include "march_hardware/error/hardware_exception.h"
