+++ march_hardware/test/mocks/MockEncoderIncremental.cpp
@@ -1,4 +1,4 @@
-#include "march_hardware/EncoderIncremental.h"
+#include "march_hardware/encoder/EncoderIncremental.h"
 
 #include <gmock/gmock.h>
